requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
PyYAML==6.0.1
requests==2.31.0
selenium==4.13.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
undetected-chromedriver==3.5.3
urllib3==2.0.6
websockets==11.0.3
wsproto==1.2.0